requirements.txt
accelerate==1.6.0
aiohttp==3.13.3
autogen==0.10.3
backoff==2.2.1
codetiming==1.4.0
datasets>=3.2.0
deepspeed>=0.15.4
einops==0.8.2
filelock==3.25.0
huggingface_hub==0.36.2
Jinja2==3.1.6
mathruler==0.1.0
mlflow==3.10.0
modelscope>=1.25.0
numpy>=2.0,<2.6
omegaconf==2.3.0
openai==2.24.0
opencv-python-headless==4.11.0.86
peft>=0.14.0
Pillow==12.1.1
psutil==7.0.0
pyarrow==22.0.0
PyYAML==6.0.2
qwen-vl-utils==0.0.14
ray[default]>=2.48.0,<3
safetensors>=0.5.3
swanlab>=0.7.8
tensordict==0.10.0
tensorboard>=2.18.0
torchdata==0.11.0
tqdm==4.67.1
transformers==4.57.6
trl>=0.18.1
wandb>=0.18.0
